pyproject.toml
[project]
name = "cphos-ai-scoring"
version = "0.1.0"
description = "CPHOS AI 自动阅卷系统 — 基于 VLM/LLM 的物理竞赛手写答卷自动评分"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pillow>=11.0.0,<12",
    "pypdfium2>=4.30.0,<5",
    "python-dotenv>=1.0.1,<2",
    "pyyaml>=6.0.1,<7",
    "matplotlib>=3.8.0,<4",
]

[project.scripts]
ai-scoring = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
